funcionalidadeDesenvolvidaParcialmenteAclicarEmMaisDeUmSaborNaoDeixaSelecionar
